Add files via upload
Quality of life updates to DynamicMarkovBlanketDiscovery and a few other routines.  The most important is added functionality for finding n objects interacting in a common environment.  This includes a .particular_assignments() routine which provides a label for each microscopic observable that identifies it as either the environment (label=0) or as part of object n (label=n).  This label includes both internal components and the boundary of a given object.