Add files via upload
Added multivariate normal with automatic relevance determination (ARD) prior 

Added a routine for Multinomial Logistic regression with the ARD prior.  This routine is a VB implementation of the polyagamma trick for use with a stick breaking construction.  It avoids the typical gibbs sampling step by making use of a factorized posterior that includes a distribution over the polyagamma augmenting variable.  In my experience this approach is as fast as taking a single gibbs sample with little cost to performance.   

This addition makes it possible to implement a non-linear dynamical system in the DMBD code using recurrent switching linear dynamical system.