NLRegression Updates
NLRegression now has dedicated low rank and full rank versions as well as a Multinomial based version.

The difference between the multinomial is that mixture components which implement the non-linearity are identified by linearly segregating the space of regressors while the low and full rank versions use Gaussian blobs to segregate the regressors.  

Also Added a dedicated ReducedRankRegression routine.  This effectively performs a Bayesian version of canonical correlation analysis to identify a low rank linear transform.  The key advantage here is the elimination of large matrix inversions when performing parameter updates and the cost of a lot of little matrix inversions when implementing the E-step.