Add files via upload
Added new features to Wishart distribution to allow one to implement constraints on posterior means of the trace or log det and reduced computational load by storing matrices in a spectral format.  This reduces memory cost and eliminates a few unnecessary matrix inversions.  This change broke a bunch of stuff so I'm uploading latest versions of everything I believe to be still working.