Removed spectral representation of covariances
For reasons unknown, the spectral representation of covariance matrices was leading to numerical instability when applied to large observation spaces.  I restored the original version but added some features to check for situations where covariance matrices are ill conditioned.  Ill conditioned covariance matrices are now hackily replaced with the prior covariance whenever they occur.  A message is generated when this happens.