Major Update
Two changes to how masking is done in the MatrixNormalWishart and MatrixNormalGamma routines. 
     X_mask removes from consideration the regressors with X_mask = False from consideration completely.  
     mask uses Lagrange multipliers to set the mean of the mask = False entries to zero.  

This change has the effect of speeding up inference by reducing the size of the matrix inversions used for learning.  It also speeds up inference for the DMBD and leads to a better approximation for the ELBO.  Note that the ELBO calculation is still approximate and may still decrease, but this now occurs only near to the point of convergence.  

I also uploaded a few new routines to support switching linear dynamical systems and deep networks of mixtures of linear transforms.  There are currently multiple forms that the linear transforms can take, the ones that work the best are dMixtureofLinearTransforms and NLRegression_Multiomial.  In principle, these two are the same, but dMixtureofLinearTransforms has a better backward routine.  

Finally, pulled apart the dmbd_test.py so that individual test cases can be more easily performed:

Newtons_Cradle_example.py
Flame_example.py
Artificial_Life_example.py
Flocking_example.py