Added Backward routine to MatrixNormalGamma
Also homogenized the code for MatrixNormalGamma and MatrixNormalWishart.  Also changed how the forward and backward routines works.  

The new logic is that forward and backward routines operate on distributions with EinvSigma() routines and perform an exact marginalization
     forward(pX)
     backward(pX)

If X or Y are tensors of observations instead of distributions Elog_like_X(Y) instead of backward and predict(X) instead of forward.  

Note that as a result of this change forward and backward no longer handle cases in which the inputs are Delta distributions.